don't discard un-applied updates, if simulation is not in progress

+++ xlive/Blam/Engine/simulation/simulation.cpp
@@ -85,8 +85,13 @@ void __cdecl simulation_update_pregame()
 
 void simulation_destroy_update()
 {
-    // remove everything from the queue
-    simulation_get_world()->destroy_update();
+    // only discard these updates if they were actually
+    // during simulation
+    if (simulation_in_progress())
+    {
+        // remove everything from the queue
+		simulation_get_world()->destroy_update();
+    }
 }
 
 void simulation_apply_patches()
